pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["ncheymbamalu <nchey.learnings@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">3.10,<3.11"
ipykernel = "^6.29.2"
jupyter = "^1.0.0"
loguru = "^0.7.2"
nbformat = "5.10.4"
numpy = "1.26.4"
omegaconf = "2.3.0"
pandas = "^2.1.1"
psycopg2-binary = "^2.9.9"
pyarrow = "17.0.0"
pydantic = "2.9.2"
pydantic-settings = "2.5.2"
python-dotenv = "^1.0.1"
scikit-learn = "^1.3.1"
sqlalchemy = "^2.0.35"
xgboost = "2.1.1"

[tool.poetry.group.dev.dependencies]
Flake8-pyproject = "1.2.3"
isort = "5.13.2"
pylint = "3.2.5"
pytest = "7.4.4"
ruff = "0.6.9"

[tool.flake8]
max-line-length = 100
import-order-style = "pep8"
application_import_names = "src"
exclude = [".git", "__pycache__"]
extend-ignore = ["E501", "W503", "W504", "E203", "I201", "I202"]
per-file-ignores = "src/database.py: W605"

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
indent = 4
lines_between_types = 1
known_first_party = "src"
default_section = "THIRDPARTY"
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]

[tool.pylint]
disable = ["W1401", "R0903"]

[tool.ruff]
line-length = 100
target-version = "py310"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
